+++ kernel/common/klib.h
@@ -11,7 +11,7 @@
 bool open(resource_descriptor_t* desc, resource_t* resource);
 bool read(resource_t res, char* buffer, int size, int* outSize);
 bool write(resource_t res, char* buffer, int size, int* outSize);
-bool seek(resource_t res, int relative, int offset);
+bool seek(resource_t res, int relative, ptrdiff_t offset);
 bool tell(resource_t res, int* offset);
 bool close(resource_t res);
